+++ crates/zng-wgt-fill/src/node.rs
@@ -1075,10 +1075,14 @@ pub fn flood(color: impl IntoVar<Rgba>) -> impl UiNode {
             }
         }
         UiNodeOp::Render { frame } => {
-            frame.push_color(PxRect::from_size(render_size), frame_key.bind_var(&color, |&c| c));
+            if !render_size.is_empty() {
+                frame.push_color(PxRect::from_size(render_size), frame_key.bind_var(&color, |&c| c));
+            }
         }
         UiNodeOp::RenderUpdate { update } => {
-            update.update_color_opt(frame_key.update_var(&color, |&c| c));
+            if !render_size.is_empty() {
+                update.update_color_opt(frame_key.update_var(&color, |&c| c));
+            }
         }
         _ => {}
     })