+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Fix warning on `flood` with 0 area.
 * Add `zng::task::Progress` value type for reporting a task progress status.
 * Add `zng::progress` with `ProgressView` widget for displaying a task progress status.